rebuild owif and enigma2 when extra-rcmodels repo is changed
Add recipe for extra_rc_models and add it as dependency to
OpenWebif and Enigma2.
This forces a rebuild of OpenWebif and Enigma2 when we add
something new to extra_rc_models.
So that at least the new extra_rc_models changes are in the image without waiting
on a new GIT SRCREV OpenWebif/Enigma2.
Users with a feed server can do a 'force --reinstall', if they
need the new extra_rc_models changes.

This installs just one tiny file in sysroots-components.
I know it's a bit weirdy to add a package like extra_rc_models as dependency.
But if someone knows a better solution with all the in and outs and side-effects
or disagree shout it out.